created new file for createGoalView
